Ask for admin name/subject when creating creating first provisioner
In enableAdmin=true mode, the first provisioner is created on startup.

This patch asks for the admin name/subject when creating the provisioner.
This is necessary if the root CA has nameConstraints and only permits
certain subjects.

Otherwise, there will be an error when using the provisioner with an
admin command (e.g. `step ca admin list`). See #962