Use `X-Request-Id` as canonical request identifier (if available)
If `X-Request-Id` is available in an HTTP request made against the
CA server, it'll be used as the identifier for the request. This
slightly changes the existing behavior, which relied on the custom
`X-Smallstep-Id` header, but usage of that header is currently not
very widespread, and `X-Request-Id` is more generally known for
the use case `X-Smallstep-Id` is used for.

`X-Smallstep-Id` is currently still considered, but it'll only be
used if `X-Request-Id` is not set.