Clean up active/passive revocation tests
Amend tests to no longer assert an error message upon active revocation
and to instead validate the success on either revocation variant,
reflecting previous changes.

In theory the "Passive" boolean seems removable as it does not add any
functionality besides being logged, but it was found some instances of
it are still needed to block active revocation for SSH certificates.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>