Fix sshagentkms
Back in "Implementation of the Prometheus endpoint (#1669)"
(dd1ff9c15b0f8ca5920dac6a0ff080b7f0cfb8be) the keyManager was wrapped in
another layer, which caused the key that ssh.Signer tries to use to be
a agent.Key, which ssh.Signer doesn't really know what to do with.

By inspecting the instrumentedKMSSigner to see if it wraps a
WrappedSSHSigner we can extract the original signer which works
with ssh.Signer and make sshagentkms work again.

Signed-off-by: Anton Lundin <glance@ac2.se>